Work on Menu 3 Attack
Setup PrintGrid() to display the player's hit or misses on the grid. (haven't tried)
Added/Fixed queries to retrieve attack data and opponent's ships coordinates.
Created an inf loop until one of the players wins, each round will have the grid displayed and the user asked where they would like to attack. (haven't tested)
Created the if statement to input into the database if the attack was either a hit or a miss